# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "轻量级音乐服务器 API 文档，支持本地音乐管理、网络歌曲、电台和歌单功能",
        "title": "MiMusic API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/mimusic-org/mimusic",
            "email": "im.hanxi@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0"
        },
        "version": "1.4.1"
    },
    "host": "localhost:58091",
    "basePath": "/api/v1",
    "paths": {
...
        "/cache-manage/clean": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除服务端所有已缓存的音乐文件,清理后需要重新下载",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "缓存管理"
                ],
                "summary": "清理全部音乐缓存",
                "responses": {
                    "200": {
                        "description": "清理成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "清理失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/cache-manage/config": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取服务端音乐缓存的配置信息,包括最大缓存大小限制",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "缓存管理"
                ],
                "summary": "获取缓存配置",
                "responses": {
                    "200": {
                        "description": "缓存配置",
                        "schema": {
                            "$ref": "#/definitions/services.CacheConfig"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新服务端音乐缓存的配置,如最大缓存大小。更新后会自动触发 LRU 淘汰检查。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "缓存管理"
                ],
                "summary": "更新缓存配置",
                "parameters": [
                    {
                        "description": "缓存配置",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.CacheConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新后的缓存配置",
                        "schema": {
                            "$ref": "#/definitions/services.CacheConfig"
                        }
                    },
                    "400": {
                        "description": "请求参数无效",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "更新失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/cache-manage/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取服务端音乐缓存的统计信息,包括总大小、文件数量和最大缓存限制",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "缓存管理"
                ],
                "summary": "获取缓存统计信息",
                "responses": {
                    "200": {
                        "description": "缓存统计信息",
                        "schema": {
                            "$ref": "#/definitions/services.CacheStats"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "services.CacheConfig": {
            "type": "object",
            "properties": {
                "max_size": {
                    "description": "最大缓存大小（字节），0 表示无限制",
                    "type": "integer"
                }
            }
        },
        "services.CacheStats": {
            "type": "object",
            "properties": {
                "file_count": {
                    "description": "文件数量",
                    "type": "integer"
                },
                "max_size": {
                    "description": "最大缓存大小（字节），0 表示无限制",
                    "type": "integer"
                },
                "total_size": {
                    "description": "总大小（字节）",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "输入 \"Bearer {token}\" 进行认证",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}